requirements.txt
Flask>=2.2
Flask-Login>=0.6
Flask-WTF>=1.1
Flask-SQLAlchemy>=2.5
Flask-RESTful>=0.3
Flask-Mail>=0.9
WTForms>=3.0
email-validator
